+++ bot/botkit/src/bot-manager.ts
@@ -38,21 +38,20 @@ const logInfo = debug('dxos:botkit');
 // File where information about running bots is stored.
 export const BOTS_DUMP_FILE = 'out/factory-state';
 
+export type BotId = string;
+
 export interface BotInfo {
-  botId: string
+  botId: BotId
   id: string
+  installDirectory: string
+  storageDirectory: string
+  spawnOptions: Spawn.SpawnOptions
   parties: string[]
   started: any
   lastActive: any
   stopped: boolean
   name: string
-  type?: any
-  childDir: string
-  command: string
-  args: string[]
   env: string
-  process: any
-  watcher?: any
 }
 
 interface Options {
@@ -98,16 +97,16 @@ export class BotManager {
     this._emitBotEvent = emitBotEvent;
 
     this._localDev = this._config.get('bot.localDev');
-    this._botsFile = path.join(process.cwd(), BOTS_DUMP_FILE);
+    this._botsFile = path.join(process.cwd(), this._config.get('bot.dumpFile', BOTS_DUMP_FILE));
 
     this._registry = new Registry(this._config.get('services.wns.server'), this._config.get('services.wns.chainId'));
 
     ensureFileSync(this._botsFile);
 
     for (const container of Object.values(this._botContainers)) {
-      container.on('bot-close', async (botId: string, code: number) => {
+      container.botExit.on(async ({ botId, exitCode }) => {
         const botInfo = this._bots.get(botId);
-        if (!code && botInfo) {
+        if (!exitCode && botInfo) {
           botInfo.stopped = true;
           await this._saveBotsToFile();
         }
@@ -168,6 +167,10 @@ export class BotManager {
 
     log(`Spawn bot request for ${botName || ipfsCID || displayName} env: ${env}`);
 
+    if (this._botContainers[env] === undefined) {
+      throw new Error(`Unknown env ${env}. Available envs are: ${Object.keys(this._botContainers)}`);
+    }
+
     assert(id, 'Invalid Bot Id.');
     assert(displayName, 'Invalid Bot Name.');
 
@@ -177,9 +180,21 @@ export class BotManager {
     const installDirectory = await this._sourceManager.downloadAndInstallBot(id, ipfsCID, options);
     assert(installDirectory, `Invalid install directory for bot: ${botName || ipfsCID}`);
 
-    const params = await this._botContainers[env].getBotAttributes(botId, installDirectory, options);
+    this._bots.set(botId, {
+      botId,
+      id,
+      installDirectory,
+      storageDirectory: path.join(process.cwd(), '.bots', botId),
+      spawnOptions: options,
+      parties: [],
+      stopped: false,
+      name,
+      env,
+      started: 0,
+      lastActive: 0
+    });
 
-    return this._startBot(botId, { botName, env, name, ...params });
+    return this._startBot(botId);
   }
 
   /**
@@ -196,7 +211,8 @@ export class BotManager {
     const botInfo = this._bots.get(botId);
     assert(botInfo, 'Invalid Bot Id');
 
-    await this._botContainers[botInfo.env].killBot(botInfo);
+    await this._botContainers[botInfo.env].stopBot(botInfo);
+    await fs.remove(botInfo.storageDirectory);
     this._bots.delete(botId);
     await this._saveBotsToFile();
 
@@ -206,7 +222,8 @@ export class BotManager {
   async killAllBots () {
     for await (const botInfo of this._bots.values()) {
       if (this._botContainers[botInfo.env]) {
-        await this._botContainers[botInfo.env].killBot(botInfo);
+        await this._botContainers[botInfo.env].stopBot(botInfo);
+        await fs.remove(botInfo.storageDirectory);
       }
     }
     this._bots.clear();
@@ -291,11 +308,12 @@ export class BotManager {
    * @param botId
    * @param options
    */
-  private async _startBot (botId: string, options: any = {}) {
-    let botInfo = this._bots.get(botId);
-    botInfo = await this._botContainers[options.env].startBot(botId, botInfo, options);
+  private async _startBot (botId: string) {
+    const botInfo = this._bots.get(botId);
+    assert(botInfo);
+    log(`_startBot ${JSON.stringify(botInfo)}`);
+    await this._botContainers[botInfo.env].startBot(botInfo);
 
-    this._bots.set(botId, botInfo!);
     await this._saveBotsToFile();
 
     return botId;
@@ -319,7 +337,7 @@ export class BotManager {
   }
 
   private async _saveBotsToFile () {
-    const data = [...this._bots.values()].map(botInfo => this._botContainers[botInfo.env].serializeBot(botInfo));
+    const data = [...this._bots.values()];
     await fs.writeJSON(this._botsFile, data);
   }