+++ bot/botkit/src/index.ts
@@ -2,8 +2,7 @@
 // Copyright 2020 DXOS.org
 //
 
-export * from './bot';
 export * from './bot-factory';
 export * from './containers';
 export * from './config';
-export { NATIVE_ENV, NODE_ENV } from './env';
+export { NATIVE_ENV, NODE_ENV, BROWSER_ENV } from './env';