+++ packages/react-appkit/src/components/PartySettingsDialog.js
@@ -44,17 +44,18 @@ const useStyles = makeStyles(theme => ({
   }
 }));
 
-const PartySettingsDialog = ({ open, onClose, properties = {}, onExport, displayName, onDisplayNameChange }) => {
+const PartySettingsDialog = ({ open, onClose, properties = {}, onExport, displayName: initialDisplayName }) => {
   const classes = useStyles();
-  const [subscribed] = useState(properties.subscribed);
+  const [active, setActive] = useState(properties.active);
   const [showDeleted, setShowDeleted] = useState(properties.showDeleted);
   const [inProgress, setInProgress] = useState(false);
   const [error, setError] = useState(undefined);
   const [exportedCid, setExportedCid] = useState(undefined);
   const [copiedSnackBarOpen, setCopiedSnackBarOpen] = useState(false);
+  const [displayName, setDisplayName] = useState(initialDisplayName);
 
   const handleClose = () => {
-    onClose({ subscribed, showDeleted, displayName });
+    onClose({ active, showDeleted, displayName });
   };
 
   const handleExportToIPFS = async () => {
@@ -83,28 +84,25 @@ const PartySettingsDialog = ({ open, onClose, properties = {}, onExport, display
       </DialogTitle>
 
       <DialogContent>
-        {false && ( // disabled until https://github.com/dxos/echo/issues/246 and are resolved https://github.com/dxos/echo/issues/248
-          <EditableText
-            label='Name'
-            value={displayName}
-            onUpdate={onDisplayNameChange}
-          />
-        )}
+        <EditableText
+          label='Name'
+          value={displayName}
+          onUpdate={setDisplayName}
+        />
 
         {/* TODO(burdon): Implement state and handlers. */}
         <FormControl className={classes.form}>
           <FormGroup>
-            {/* Not implemented for the new echo */}
-            {/* <FormControlLabel
+            <FormControlLabel
               control={
                 <Checkbox
-                  checked={subscribed}
-                  value={subscribed}
-                  onChange={() => setSubscribed(!subscribed)}
+                  checked={active}
+                  value={active}
+                  onChange={() => setActive(!active)}
                 />
               }
               label='Active'
-            /> */}
+            />
             <FormControlLabel
               control={
                 <Checkbox