+++ packages/react-appkit/src/containers/AppBar.js
@@ -21,12 +21,11 @@ import ShareIcon from '@material-ui/icons/Share';
 
 import { BotFactoryClient } from '@dxos/botkit-client';
 import { generatePasscode } from '@dxos/credentials';
-import { encrypt, decrypt, keyToBuffer, verify, SIGNATURE_LENGTH, humanize } from '@dxos/crypto';
+import { encrypt, keyToBuffer, verify, SIGNATURE_LENGTH } from '@dxos/crypto';
 import { useClient, useConfig, useProfile } from '@dxos/react-client';
 
 import BotDialog from '../components/BotDialog';
 import ExportKeyringDialog from '../components/ExportKeyringDialog';
-import ImportKeyringDialog from '../components/ImportKeyringDialog';
 import InvitationDialog from '../components/InvitationDialog';
 import { Action, useActionHandler, useAppRouter } from '../hooks';
 
@@ -38,7 +37,6 @@ const ACTION_USER_INVITATION = 1;
 const ACTION_DEVICE_INVITATION = 2;
 const ACTION_BOT_INVITATION = 3;
 const ACTION_EXPORT_KEYRING = 4;
-const ACTION_IMPORT_KEYRING = 5;
 const ACTION_RESET_STORAGE = 6;
 const ACTION_OPEN_SETTINGS = 7;
 const ACTION_OPEN_PARTY_HOME = 8;
@@ -169,17 +167,17 @@ const AppBar = ({
    * Initiate the device invitation flow.
    */
   const handleDeviceInvite = async () => {
-    const invitation = await client.partyManager.identityManager.deviceManager.addDevice(
-      (invitation, secret) => secret && secret.equals(invitation.secret),
-      () => {
-        const passcode = generatePasscode();
-        setPasscode(passcode);
-        return Buffer.from(passcode);
-      },
-      {
-        onFinish: () => setDialog()
-      }
-    );
+    const secretValidator = (invitation, secret) => secret && secret.equals(invitation.secret);
+    const secretProvider = () => {
+      const passcode = generatePasscode();
+      setPasscode(passcode);
+      return Buffer.from(passcode);
+    };
+    const onFinish = () => setDialog();
+
+    // TODO(rzadp): Uncomment after updating ECHO.
+    // const invitation = await client.createHaloInvitation({ secretProvider, secretValidator }, { onFinish });
+    const invitation = await client.echo._identityManager.halo.invitationManager.createInvitation({ secretProvider, secretValidator }, { onFinish });
 
     setInvitation(invitation);
     setPasscode(null);
@@ -187,16 +185,8 @@ const AppBar = ({
     setDialog({ dialog: ACTION_DEVICE_INVITATION });
   };
 
-  // TODO(burdon): Broken.
-  const keyringEncrypter = async passphrase => {
-    const keyring = await client.keyringStore.getKeyring(topic);
-    return encrypt(keyring.toJSON(), passphrase);
-  };
-
-  // TODO(burdon): Broken.
-  const keyringDecrypter = async (data, passphrase) => {
-    const keyring = await client.keyringStore.getKeyring(topic);
-    await keyring.loadJSON(decrypt(data, passphrase));
+  const keyringEncrypter = passphrase => {
+    return encrypt(client.echo.keyring.toJSON(), passphrase);
   };
 
   //
@@ -225,13 +215,6 @@ const AppBar = ({
       label: 'Export keys',
       handler: () => {
         setDialog({ dialog: ACTION_EXPORT_KEYRING });
-      }
-    },
-
-    [ACTION_IMPORT_KEYRING]: {
-      label: 'Import keys',
-      handler: () => {
-        setDialog({ dialog: ACTION_IMPORT_KEYRING });
       }
     },
 
@@ -294,13 +277,10 @@ const AppBar = ({
   //
 
   const menuItems = [
-    // action(ACTION_DEVICE_INVITATION)
+    action(ACTION_DEVICE_INVITATION)
   ];
 
-  // if (topic) {
-  //   menuItems.push(action(ACTION_EXPORT_KEYRING));
-  //   menuItems.push(action(ACTION_IMPORT_KEYRING));
-  // }
+  // menuItems.push(action(ACTION_EXPORT_KEYRING)); // ISSUE: https://github.com/dxos/echo/issues/339#issuecomment-735918728
 
   if (onSettingsOpened) {
     menuItems.push(action(ACTION_OPEN_SETTINGS));
@@ -376,16 +356,6 @@ const AppBar = ({
           encrypter={keyringEncrypter}
         />
       )
-    },
-    {
-      key: ACTION_IMPORT_KEYRING,
-      dialog: (
-        <ImportKeyringDialog
-          open={dialog === ACTION_IMPORT_KEYRING}
-          onClose={handleClose}
-          decrypter={keyringDecrypter}
-        />
-      )
     }
   ];
 
@@ -432,7 +402,7 @@ const AppBar = ({
         ))}
 
         <div>
-          <Tooltip title={profile.username || humanize(profile.publicKey)}>
+          <Tooltip title={profile.username || 'Loading...'}>
             <IconButton color='inherit'>
               <ProfileIcon />
             </IconButton>