+++ packages/react-appkit/src/containers/DefaultItemList.js
@@ -10,7 +10,7 @@ import { Divider } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
 import TreeView from '@material-ui/lab/TreeView';
 
-import { keyToBuffer } from '@dxos/crypto';
+import { keyToBuffer, keyToString, PublicKey } from '@dxos/crypto';
 import { ObjectModel } from '@dxos/object-model';
 import { useParty, useItems } from '@dxos/react-client';
 
@@ -46,7 +46,7 @@ const useStyles = makeStyles(theme => ({
 const DefaultItemList = () => {
   const router = useAppRouter();
   const { topic, item: active } = useParams();
-  const party = useParty(keyToBuffer(topic));
+  const party = useParty(PublicKey.from(topic));
   const classes = useStyles();
   const [pads] = usePads();
   const items = useItems({ partyKey: keyToBuffer(topic), type: pads.map(pad => pad.type) });
@@ -54,7 +54,7 @@ const DefaultItemList = () => {
   const anchor = useRef();
 
   const handleSelect = (itemId) => {
-    router.push({ topic, item: itemId });
+    router.push({ topic: keyToString(party.key.asUint8Array()), item: itemId });
   };
 
   const handleCreate = async (type) => {